// assets/script/protobuf/holdem/code_pb.d.ts
export interface CodeMap {
// ...
  MSG_D_GD_ENTER_ROOM: 4001;
  MSG_D_GD_SEATED: 4002;
  MSG_D_GD_BRING_IN: 4003;
  MSG_D_GD_READY_START: 4004;
  MSG_D_GD_WINRATE_ADD: 4005;
  MSG_D_GD_TRIBUTE_GIVE: 4006;
  MSG_D_GD_TRIBUTE_RETURN: 4007;
  MSG_D_GD_ACTION: 4008;
  MSG_D_GD_AUTO_OP_ACTIVE: 4009;
  MSG_D_GD_STANDUP_ACTIVE: 4010;
  MSG_D_GD_LEAVE: 4011;
  MSG_D_GD_KEEP_SEAT_ACTIVE: 4012;
  MSG_D_GD_ADD_TIME: 4013;
  MSG_D_GD_BROADCAST_MSG: 4014;
  MSG_D_GD_PRIVATE_MSG: 4015;
  MSG_D_GD_ROOMERS: 4016;
  MSG_D_GD_SET_AUTO_ON_TABLE: 4017;
  MSG_D_GD_OBSERVERS: 4018;
  MSG_D_GD_SYNC_ENTER: 4019;
  MSG_D_GD_REPLAY: 4020;
  MSG_D_GD_CHAT_MEMBERS: 4021;
  MSG_S_GD_SEATED_OTHERS: 4101;
  MSG_S_GD_START_INFO: 4102;
  MSG_S_GD_WINRATE_ADD_COMPLETE: 4103;
  MSG_S_GD_TRIBUTE_GIVE_COMPLETE: 4104;
  MSG_S_GD_TRIBUTE_RETURN_COMPLETE: 4105;
  MSG_S_GD_CHIPS_CHANGE: 4106;
  MSG_S_GD_ACTION_ALL: 4107;
  MSG_S_GD_AUTO_OP: 4108;
  MSG_S_GD_STANDUP: 4109;
  MSG_S_GD_KEEP_SEAT: 4110;
  MSG_S_GD_RESULT: 4111;
  MSG_S_GD_ADD_TIME_OTHERS: 4112;
  MSG_S_GD_LEAVE_NOTIFICATION: 4113;
  MSG_S_GD_BRING_IN_FAIL: 4114;
  MSG_S_GD_HAND_CLEAR: 4115;
  MSG_S_GD_GET_MSG: 4116;
  MSG_S_GD_READY_START_ALL: 4117;
  MSG_S_GD_WAIT_READY_START: 4118;
  MSG_S_GD_VIDEO_MASK_CHANGE: 4119;
}

export const Code: CodeMap;